[UPDATE] cleaned up circular refs from repr
