+++ pyhydrate/__init__.py
@@ -1,7 +1,4 @@
 from .notation.notation_primitive import NotationPrimitive
-from .notation.notation_representation import (
-    NotationRepresentation,
-)
 from .notation.notation_structures import NotationArray, NotationObject
 from .pyhydrate import PyHydrate
 
@@ -9,6 +6,5 @@ __all__ = [
     "NotationArray",
     "NotationObject",
     "NotationPrimitive",
-    "NotationRepresentation",
     "PyHydrate",
 ]
