+++ pyhydrate/notation/__init__.py
@@ -1,10 +1,8 @@
 from .notation_primitive import NotationPrimitive
-from .notation_representation import NotationRepresentation
 from .notation_structures import NotationArray, NotationObject
 
 __all__ = [
     "NotationArray",
     "NotationObject",
     "NotationPrimitive",
-    "NotationRepresentation",
 ]
